backend/sub2/backend/requirements.txt
astroid==2.3.3
beautifulsoup4==4.9.0
branca==0.3.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9.2
cycler==0.10.0
Django==2.2.7
django-rest-swagger==2.2.0
djangorestframework==3.10.3
drf-yasg==1.17.1
flex==6.14.1
folium==0.10.1
google-images-download==2.8.0
idna==2.8
inflection==0.4.0
isort==4.3.21
itypes==1.2.0
Jinja2==2.10.3
joblib==0.14.0
jsonpointer==2.0
jwt==1.0.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.1.2
mccabe==0.6.1
mysqlclient==1.4.6
numpy==1.17.3
openapi-codec==1.3.2
packaging==20.3
pandas==0.25.3
pycparser==2.20
pylint==2.4.4
pyparsing==2.4.5
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
requests==2.22.0
rfc3987==1.3.8
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
scikit-learn==0.21.3
scikit-surprise==1.1.0
scipy==1.3.1
seaborn==0.9.0
selenium==3.141.0
simplejson==3.17.0
six==1.13.0
soupsieve==2.0
sqlparse==0.3.0
strict-rfc3339==0.7
surprise==0.1
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.7
validate-email==1.3
wrapt==1.11.2
django-cors-headers